WIP: pluggable upload destinations
Since uploads.im was going away I wanted to store my images on S3. I'm sure
that's not what most people want to do, so I abstracted it a little. This
strategy would be a good way to go for sending messages (eg twilio, gvoice).

Marking as a WIP so you all can see it, and because it's a WIP. I'm still testing, and I need to add options to the config panel.